+++ JobTracker-master/src/main/java/org/example/service/JobService.java
@@ -15,7 +15,6 @@ public class JobService {
     @Autowired
     private JobRepository jobRepository;
 
-    // Save or update a job
     public Job saveJob(Job job) {
         return jobRepository.save(job);
     }
@@ -27,12 +26,10 @@ public class JobService {
 
     // Search for a job by title
     public List<Job> findByTitle(String title) {
-        // Assuming you add a method in JobRepository to find by title
         return jobRepository.findByTitleIgnoreCase(title);
     }
 
-    // Delete a job by id
-    public void deleteJobById(int id) {
+    public void deleteJobById(Long id) {
         jobRepository.deleteById(id);
     }
